+++ CHANGES.rst
@@ -6,9 +6,13 @@ Changes
 
 - Fix @@zodbbrowser_history when using ZEO (AttributeError:
   'TransactionIterator' object has no attribute 'close', GH #31).
+
 - Fix OOBTree history browsing that regressed in 0.11.0: it was showing only a
   subset of changes (those that touched the OOBTree itself) and wasn't showing
   others (those that touched deeper OOBTreeBuckets).
+
+- Fix TypeError for SampleContainer subclasses that use a non-persistent
+  ``__data`` attribute (GH: #18) 
 
 
 0.16.2 (2020-11-24)